Extract tables to component first step
