+++ package.json
@@ -27,6 +27,7 @@
     "jsonapi-serializer": "^3.5.6",
     "koa": "^2.5.2",
     "koa-bodyparser": "^4.2.1",
+    "koa-logger": "^3.2.0",
     "koa-router": "^7.4.0",
     "mongoose": "^5.2.11",
     "nodemon": "^1.18.4"
