+++ src/components/tables/actions/findAll.js
@@ -1,0 +1,10 @@
+const Table = require('../Table');
+const tablesSerializer = require('../tables-serializer');
+
+module.exports = async (ctx, next) => {
+  const tables = await Table.find({});
+  const payload = tablesSerializer.serialize(tables);
+  ctx.body = payload
+
+  await next();
+};
