+++ src/components/tables/actions/index.js
@@ -1,0 +1,7 @@
+const findAll = require('./findAll');
+const create = require('./create');
+
+module.exports = {
+  create,
+  findAll
+};
