+++ src/components/tables/index.js
@@ -1,0 +1,5 @@
+const tableRouter = require('./router');
+
+module.exports =  {
+  routes: tableRouter.routes()
+}
